AT-1625 Updated button action configuration on the app preview page.

+++ src/app/pages/applications/app-detail/button-action/button-action.component.ts
@@ -1,13 +1,5 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output, TemplateRef, ViewChild } from '@angular/core';
 import {
-    ButtonAction,
-    DownloadButtonAction,
-    FormButtonAction,
-    OwnershipButtonAction,
-    PurchaseButtonAction,
-    ViewData
-} from './models/button-action.model';
-import {
     AppFormService,
     AuthHolderService,
     FileUploadDownloadService,
@@ -24,8 +16,11 @@ import { FullAppData, OcButtonComponent, OcFormModalComponent } from '@openchann
 import { ActivatedRoute, Router } from '@angular/router';
 import { get } from 'lodash';
 import { HttpHeaders } from '@angular/common/http';
-
-declare type ActionType = 'OWNED' | 'UNOWNED';
+import {
+    ActionButton,
+    DownloadButtonType,
+    FormButtonType
+} from '../../../../../assets/data/configData';
 
 @Component({
     selector: 'app-button-action',
@@ -33,16 +28,12 @@ declare type ActionType = 'OWNED' | 'UNOWNED';
     styleUrls: ['./button-action.component.scss'],
 })
 export class ButtonActionComponent implements OnInit, OnDestroy {
-    @Input() buttonAction: ButtonAction;
+    @Input() buttonAction: ActionButton;
     @Input() appData: FullAppData;
     @Output() readonly updateAppData: EventEmitter<void> = new EventEmitter<void>();
 
     @ViewChild('rejectButton') rejectButton: TemplateRef<OcButtonComponent>;
     @ViewChild('confirmButton') confirmButton: TemplateRef<OcButtonComponent>;
-
-    viewData: ViewData;
-
-    actionType: ActionType = null;
 
     private $destroy: Subject<void> = new Subject();
 
@@ -65,7 +56,6 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
 
     ngOnInit(): void {
         this.loader = this.loadingBar.useRef();
-        this.initButtonData();
     }
 
     ngOnDestroy(): void {
@@ -77,60 +67,28 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
         this.modal.dismissAll();
     }
 
-    initButtonData(): void {
-        switch (this.buttonAction.type) {
-            case 'form':
-                this.setViewData(null, this.buttonAction as FormButtonAction);
-                break;
-            case 'install':
-                if (this.appData?.ownership?.ownershipStatus === 'active') {
-                    this.setViewData('OWNED', (this.buttonAction as OwnershipButtonAction)?.owned);
-                } else {
-                    this.setViewData('UNOWNED', (this.buttonAction as OwnershipButtonAction)?.unowned);
-                }
-                break;
-            case 'download':
-                this.viewData = {
-                    button: (this.buttonAction as DownloadButtonAction).button,
-                    message: null,
-                };
-                break;
-            case 'purchase':
-                this.viewData = {
-                    button: (this.buttonAction as PurchaseButtonAction).button,
-                    message: null,
-                };
-                break;
-            default:
-                this.toasterService.error(`Error: invalid button type: ${this.buttonAction.type}`);
-        }
-    }
 
     onClick(): void {
         switch (this.buttonAction.type) {
             case 'form':
-                this.processForm(this.buttonAction as FormButtonAction);
+                this.processForm(this.buttonAction);
                 break;
             case 'install':
-                this.processOwnership();
+                this.installOwnership();
+                break;
+            case 'uninstall':
+                this.uninstallOwnership();
                 break;
             case 'download':
-                this.downloadFile(this.buttonAction as DownloadButtonAction);
+                this.downloadFile(this.buttonAction);
                 break;
             case 'purchase':
                 this.processPurchase();
                 break;
-            default:
-                this.toasterService.error(`Error: invalid button type: ${this.buttonAction.type}`);
         }
     }
 
-    private setViewData(actionType: ActionType, viewData: ViewData): void {
-        this.actionType = actionType;
-        this.viewData = viewData;
-    }
-
-    private processForm(formAction: FormButtonAction): void {
+    private processForm(formAction: FormButtonType): void {
         if (!this.inProcess) {
             this.loader.start();
             this.inProcess = true;
@@ -162,7 +120,7 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
                         });
                     },
                     () => {
-                        this.toasterService.error(formAction.message.notFound);
+                        this.toasterService.error(formAction.showToaster.notFoundFormMessage);
                         this.loader.complete();
                         this.inProcess = false;
                     },
@@ -170,25 +128,10 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
         }
     }
 
-    private processOwnership(): void {
-        if (this.authService.isLoggedInUser()) {
-            switch (this.actionType) {
-                case 'OWNED':
-                    this.uninstallOwnership();
-                    break;
-                case 'UNOWNED':
-                    this.installOwnership();
-                    break;
-                default:
-                    this.toasterService.error(`Error: invalid owned button type: ${this.actionType}`);
-            }
-        } else {
-            this.router.navigate(['login'], { queryParams: { returnUrl: window.location.pathname } }).then();
-        }
-    }
-
     private installOwnership(): void {
-        if (this.appData?.model?.length > 0) {
+        if (!this.authService.isLoggedInUser()) {
+            this.navigateToLoginPage();
+        } else if (this.appData?.model?.length > 0) {
             this.processAction(
                 this.ownershipService.installOwnership(
                     {
@@ -205,11 +148,15 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
     }
 
     private uninstallOwnership(): void {
-        this.processAction(
-            this.ownershipService.uninstallOwnership(this.appData.ownership.ownershipId, new HttpHeaders({ 'x-handle-error': '403, 500' })),
-            error => this.handleOwnershipResponseError(error, 'You don’t have permission to uninstall this app'),
-            false,
-        );
+        if (!this.authService.isLoggedInUser()) {
+            this.navigateToLoginPage();
+        } else {
+            this.processAction(
+                this.ownershipService.uninstallOwnership(this.appData.ownership.ownershipId, new HttpHeaders({'x-handle-error': '403, 500'})),
+                error => this.handleOwnershipResponseError(error, 'You don’t have permission to uninstall this app'),
+                false,
+            );
+        }
     }
 
     private processAction<T>(action: Observable<T>, errorHandler?: (error: any) => Observable<any>, reportStatistic: boolean = true): void {
@@ -228,16 +175,16 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
                             ? errorHandler
                             : error => {
                                   this.inProcess = false;
-                                  if (error.status !== 429 && this.viewData?.message?.fail) {
-                                      this.toasterService.error(this.viewData?.message?.fail);
+                                  if (error.status !== 429 && this.buttonAction?.showToaster?.errorMessage) {
+                                      this.toasterService.error(this.buttonAction?.showToaster?.errorMessage);
                                   }
                                   return throwError(error);
                               },
                     ),
                     tap(() => {
                         this.inProcess = false;
-                        if (this.viewData?.message?.success) {
-                            this.toasterService.success(this.viewData?.message?.success);
+                        if (this.buttonAction?.showToaster?.successMessage) {
+                            this.toasterService.success(this.buttonAction?.showToaster?.successMessage);
                         }
                         this.updateAppData.emit();
                     }),
@@ -262,8 +209,8 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
         }
     }
 
-    private downloadFile(actionConfig: DownloadButtonAction): void {
-        const file = get(this.appData, actionConfig.fileField);
+    private downloadFile(actionConfig: DownloadButtonType): void {
+        const file = get(this.appData, actionConfig.pathToFile);
         const regex: RegExp = new RegExp(/^(http(s)?:)?\/\//gm);
         if (regex.test(file)) {
             window.open(file);
@@ -288,8 +235,8 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
         this.inProcess = false;
         if (error.status === 403) {
             this.toasterService.error(forbiddenMessage);
-        } else if (this.viewData?.message?.fail) {
-            this.toasterService.error(this.viewData?.message?.fail);
+        } else if (this.buttonAction?.showToaster?.errorMessage) {
+            this.toasterService.error(this.buttonAction?.showToaster?.errorMessage);
         }
         return throwError(error);
     }
@@ -297,5 +244,9 @@ export class ButtonActionComponent implements OnInit, OnDestroy {
     private processPurchase(): void {
         const safeName = this.activeRoute.snapshot.paramMap.get('safeName');
         this.router.navigate(['/checkout', safeName]).then();
+    }
+
+    private navigateToLoginPage(): void {
+        this.router.navigate(['login'], { queryParams: { returnUrl: window.location.pathname } }).then();
     }
 }
